Creating new Fastapi version
